+++ src/main/java/com/jexis/jexis_backend/wallet/application/useCases/SyncBalanceUseCase.java
@@ -21,7 +21,7 @@ public class SyncBalanceUseCase {
 
         Wallet wallet = getWalletByFAIdUseCase.execute(financialAccount.getId());
 
-        wallet.setAvailableBalance(BigDecimal.valueOf(financialAccount.getBalance().getCash().get("usd")));
+        wallet.setAvailableBalance(financialAccount.getBalance().getCash().get("usd").intValue());
         repo.save(wallet);
     }
 }
